refactor(releases): separate tables and modals
